django_celery/requirements.txt
celery==5.0
django-celery-beat==2.0
django-celery-results==2.0